complete update of the data splits
