+++ notebooks/1.data_splits/1.data_splits.sh
@@ -7,5 +7,4 @@ conda activate cell-injury
 jupyter nbconvert --to python --output-dir=nbconverted/ *.ipynb
 
 # run the scripts
-python nbconverted/1.1_cell_injury_data_splits.py
-python nbconverted/1.2_aligned_cell_injury_data_splits.py
+python nbconverted/1.data_splits.py
