+++ results/1.data_splits/aligned_summary_data_split.csv
@@ -1,0 +1,16 @@
+Cellular Injury,Number of Wells (Total Data),Number of Wells (Train Split),Number of Wells (Test Split),Number of Wells (Plate Holdout),Number of Wells (Treatment Holdout),Number of Wells (Well Holdout)
+Control,9855,6726,1682,1072,0,375
+Cytoskeletal,1472,882,220,181,12,177
+Miscellaneous,1302,805,201,171,18,107
+Kinase,1104,600,150,120,12,222
+Genotoxin,944,590,148,73,48,85
+Hsp90,552,334,84,54,0,80
+Redox,312,172,43,54,24,19
+Saponin,288,130,33,102,12,11
+HDAC,168,110,28,30,0,0
+Mitochondria,144,92,23,12,0,17
+Proteasome,144,94,23,24,0,3
+Nonspecific reactive,128,84,21,19,0,4
+Ferroptosis,96,66,16,6,0,8
+Tannin,96,60,15,18,0,3
+mTOR,96,56,14,12,0,14
